refactor(speech): graduate STT/TTS providers to top-level src/speech/
Move providers, types, and fakes out of channels/voice/speech so speech is
infrastructure every channel depends down into. Rename speech-media-service
to media-service and add a public barrel. Import-path rewrites only.
