+++ src/index.ts
@@ -83,9 +83,7 @@ import { VoiceSessionBridge } from './channels/voice/session-bridge.js';
 import { VoiceSessionStore } from './channels/voice/session-store.js';
 import { VoiceRuntime } from './channels/voice/voice-runtime.js';
 import { evaluateVoiceModelCapabilities } from './channels/voice/voice-model-capabilities.js';
-import { DeepgramSttProvider } from './channels/voice/speech/deepgram-stt.js';
-import { CartesiaTtsProvider } from './channels/voice/speech/cartesia-tts.js';
-import { SpeechMediaService } from './speech/speech-media-service.js';
+import { DeepgramSttProvider, CartesiaTtsProvider, SpeechMediaService } from './speech/index.js';
 import { LiveKitRoomSession } from './channels/voice/livekit/room-session.js';
 import { deleteVoiceRoom, listVoiceRooms } from './channels/voice/livekit/token.js';
 import { loadAuthConfig } from './contacts/config-loader.js';
