+++ src/speech/cartesia-tts.ts
@@ -1,4 +1,4 @@
-import type { Logger } from '../../../logger.js';
+import type { Logger } from '../logger.js';
 import {
   TtsHttpError,
   type AudioFileFormat,
