+++ src/speech/deepgram-stt.ts
@@ -1,4 +1,4 @@
-import type { Logger } from '../../../logger.js';
+import type { Logger } from '../logger.js';
 import {
   SttHttpError,
   type BatchSpeechToTextProvider,
