+++ src/speech/fake-stt.test.ts
