+++ src/speech/fake-stt.ts
